display_app/requirements.txt
Pillow
numpy
RPi.GPIO
spidev
flask
epd-library